requirements.txt
torch>=2.2.0,<2.5
torchvision>=0.17.0,<0.20
segmentation-models-pytorch>=0.3.3,<0.4
xgboost>=2.0.0,<2.2
shap>=0.44.0,<0.46
rasterio>=1.3.9,<1.5
GDAL>=3.8.0,<3.10
pyproj>=3.6.0,<3.8
numpy>=1.26.0,<2.1
pandas>=2.1.0,<2.3
scikit-learn>=1.4.0,<1.6
scipy>=1.12.0,<1.15
matplotlib>=3.8.0,<3.10
seaborn>=0.13.0,<0.14
albumentations>=1.3.1,<1.5
opencv-python>=4.9.0,<4.11
pyyaml>=6.0,<6.1
tqdm>=4.66.0,<4.70
pytest>=8.0.0,<8.4
kagglehub>=0.2.0,<0.4
